test_requirements.txt
# Test dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
Pillow==10.0.1
requests==2.31.0
responses==0.24.0